refactor(events): split events into separate files
