+++ src/CMakeLists.txt
@@ -87,7 +87,6 @@ set(tfs_HDR
 	${CMAKE_CURRENT_LIST_DIR}/depotchest.h
 	${CMAKE_CURRENT_LIST_DIR}/depotlocker.h
 	${CMAKE_CURRENT_LIST_DIR}/enums.h
-	${CMAKE_CURRENT_LIST_DIR}/events.h
 	${CMAKE_CURRENT_LIST_DIR}/fileloader.h
 	${CMAKE_CURRENT_LIST_DIR}/game.h
 	${CMAKE_CURRENT_LIST_DIR}/groups.h
@@ -168,6 +167,8 @@ endif ()
 
 add_subdirectory(lua)
 target_sources(tfslib PRIVATE $<TARGET_OBJECTS:luascript>)
+
+add_subdirectory(events)
 
 add_custom_target(format COMMAND /usr/bin/clang-format -style=file -i ${tfs_HDR} ${tfs_SRC} ${tfs_MAIN})
