+++ src/events/events.h
@@ -1,0 +1,19 @@
+// Copyright 2023 The Forgotten Server Authors. All rights reserved.
+// Use of this source code is governed by the GPL-2.0 License that can be found in the LICENSE file.
+
+#pragma once
+
+enum class EventInfoId
+{
+	// Creature
+	CREATURE_ONHEAR,
+
+	// Monster
+	MONSTER_ONSPAWN
+};
+
+namespace tfs::events {
+
+int32_t getScriptId(EventInfoId eventInfoId);
+
+} // namespace tfs::events
