+++ src/game.cpp
@@ -9,7 +9,13 @@
 #include "configmanager.h"
 #include "creature.h"
 #include "databasetasks.h"
-#include "events.h"
+#include "events/creature.h"
+#include "events/events.h"
+#include "events/game.h"
+#include "events/monster.h"
+#include "events/party.h"
+#include "events/player.h"
+#include "globalevent.h"
 #include "housetile.h"
 #include "http/http.h"
 #include "iologindata.h"
