+++ src/lua/modules/monster.cpp
@@ -2,7 +2,7 @@
 
 #include "../../monster.h"
 
-#include "../../events.h"
+#include "../../events/events.h"
 #include "../../game.h"
 #include "../api.h"
 #include "../env.h"
