+++ src/monster.cpp
@@ -7,7 +7,7 @@
 
 #include "condition.h"
 #include "configmanager.h"
-#include "events.h"
+#include "events/monster.h"
 #include "game.h"
 #include "lua/env.h"
 #include "lua/meta.h"
