+++ src/party.cpp
@@ -6,7 +6,7 @@
 #include "party.h"
 
 #include "configmanager.h"
-#include "events.h"
+#include "events/party.h"
 #include "game.h"
 
 extern Game g_game;
