+++ src/signals.cpp
@@ -8,7 +8,8 @@
 #include "actions.h"
 #include "configmanager.h"
 #include "databasetasks.h"
-#include "events.h"
+#include "events/events.h"
+#include "events/game.h"
 #include "game.h"
 #include "monsters.h"
 #include "movement.h"
